@@ -44,48 +61,72 @@
   "sandbox": {
     "url": "https://majifix-service-group.herokuapp.com/v1/servicegroups"
   },
-  "dependencies": {
-    "@codetanzania/majifix-common": ">=0.12.4",
-    "@codetanzania/majifix-jurisdiction": ">=1.5.5",
-    "@codetanzania/majifix-priority": ">=1.4.5",
-    "@lykmapipo/express-common": ">=0.17.3",
-    "async": ">=3.0.0",
-    "express": ">=4.16.4",
-    "lodash": ">=4.17.11",
-    "moment": ">=2.24.0",
-    "mongoose-geojson-schemas": ">=0.10.5",
-    "mongoose-locale-schema": ">=0.3.1",
-    "mongoose-rest-actions": ">=0.28.1",
-    "randomcolor": ">=0.5.4"
-  },
   "devDependencies": {
-    "@benmaruchu/faker": "^4.2.1",
-    "@commitlint/cli": "^7.6.1",
-    "@commitlint/config-conventional": "^7.6.0",
-    "@commitlint/travis-cli": "^7.6.1",
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/node": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/register": "^7.5.5",
+    "@commitlint/cli": "^8.1.0",
+    "@commitlint/config-conventional": "^8.1.0",
+    "@commitlint/travis-cli": "^8.1.0",
     "chai": "^4.2.0",
-    "commitizen": "^3.1.1",
-    "cz-conventional-changelog": "^2.1.0",
-    "express": "^4.17.0",
-    "generate-changelog": "^1.7.1",
-    "grunt": "^1.0.4",
-    "grunt-apidoc": "^0.11.0",
-    "grunt-contrib-jshint": "^2.1.0",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-mocha-test": "^0.13.3",
-    "husky": "^2.3.0",
-    "jshint-stylish": "^2.2.1",
-    "lint-staged": "^8.1.7",
-    "mocha": "^6.1.4",
-    "mongoose": ">=5.5.10",
-    "prettier": "^1.17.1",
-    "sinon": "^7.3.2",
-    "sinon-chai": "^3.3.0",
-    "sinon-mongoose": "^2.3.0",
-    "supertest": "^4.0.2"
+    "commitizen": "^4.0.3",
+    "coveralls": "^3.0.6",
+    "cz-conventional-changelog": "^3.0.2",
+    "doxdox": "^3.0.0",
+    "eslint": "^6.2.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-prettier": "^6.1.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsdoc": "^15.8.1",
+    "eslint-plugin-mocha": "^6.0.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "generate-changelog": "^1.8.0",
+    "handlebars": "^4.1.2",
+    "husky": "^3.0.4",
+    "lint-staged": "^9.2.3",
+    "lodash.template": "^4.5.0",
+    "mocha": "^6.2.0",
+    "npm-check-updates": "^3.1.21",
+    "nyc": "^14.1.1",
+    "prettier": "^1.18.2",
+    "rimraf": "^3.0.0",
+    "rollup": "^1.19.4",
+    "shelljs": "^0.8.3",
+    "@benmaruchu/faker": "4.3.0",
+    "@lykmapipo/mongoose-test-helpers": "0.7.17",
+    "mongoose": "5.6.9",
+    "sinon": "7.4.1",
+    "sinon-chai": "3.3.0",
+    "sinon-mongoose": "2.3.0",
+    "supertest": "4.0.2",
+    "winston": "^3.2.1"
   },
   "peerDependencies": {
-    "mongoose": ">=5.5.10"
+    "mongoose": ">=5.6.9"
+  },
+  "dependencies": {
+    "@codetanzania/majifix-common": ">=0.15.0",
+    "@codetanzania/majifix-jurisdiction": ">=1.7.0",
+    "@codetanzania/majifix-priority": ">=1.5.0",
+    "@lykmapipo/common": ">=0.25.1",
+    "@lykmapipo/env": ">=0.12.1",
+    "@lykmapipo/express-common": ">=0.17.10",
+    "@lykmapipo/express-rest-actions": ">=0.6.7",
+    "@lykmapipo/mongoose-common": ">=0.25.10",
+    "@lykmapipo/mongoose-exportable": ">=0.3.7",
+    "async": ">=3.1.0",
+    "express": ">=4.17.1",
+    "lodash": ">=4.17.15",
+    "moment": ">=2.24.0",
+    "mongoose-geojson-schemas": ">=0.10.9",
+    "mongoose-locale-schema": ">=0.4.0",
+    "mongoose-rest-actions": ">=0.28.7"
+  },
+  "engines": {
+    "node": ">=12.4.0",
+    "npm": ">=6.9.0"
   },
   "config": {
     "commitizen": {
