@@ -1,20 +1,8 @@
-'use strict';
-
-
-/* dependencies */
-const path = require('path');
-const { expect } = require('chai');
-
-
-/* declarations */
-const ServiceGroup =
-  require(path.join(__dirname, '..', '..', 'lib', 'servicegroup.model'));
-
+import { expect } from '@lykmapipo/mongoose-test-helpers';
+import ServiceGroup from '../../src/servicegroup.model';
 
 describe('ServiceGroup', () => {
-
   describe('Statics', () => {
-
     it('should expose model name as constant', () => {
       expect(ServiceGroup.MODEL_NAME).to.exist;
       expect(ServiceGroup.MODEL_NAME).to.be.equal('ServiceGroup');
@@ -22,18 +10,19 @@ describe('ServiceGroup', () => {
 
     it('should expose autopulate as options', () => {
       expect(ServiceGroup.OPTION_AUTOPOPULATE).to.exist;
-      expect(ServiceGroup.OPTION_AUTOPOPULATE)
-        .to.be.eql({
-          select: { code: 1, name: 1, color: 1 },
-          maxDepth: 1
-        });
+      expect(ServiceGroup.OPTION_AUTOPOPULATE).to.be.eql({
+        select: { code: 1, name: 1, color: 1 },
+        maxDepth: 1,
+      });
     });
 
-    it('should expose default locale `en` when not set', () => {
-      expect(ServiceGroup.DEFAULT_LOCALE).to.exist;
-      expect(ServiceGroup.DEFAULT_LOCALE).to.equal('en');
+    it('should expose field select option', () => {
+      expect(ServiceGroup.OPTION_SELECT).to.exist;
+      expect(ServiceGroup.OPTION_SELECT).to.be.eql({
+        code: 1,
+        name: 1,
+        color: 1,
+      });
     });
-
   });
-
 });
